requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
pandas==0.24.2
pytorch==1.6.0
numpy==1.19.2
torchvision==0.7.0
